Major API rewrite
The Service class now implements a minimal API for executing Hooks. All convenience functions have to be implemented additionally. For example, for deduplication of events with the same value, or registration of multiple events listeners. An error no longer leads to a stop of the service. In general, a service can no longer be stopped. If the Hook can handle subsequent errors, you can continue to use the service. But there is a method to dispose all effects.