Fundamental API change
To allow more flexible use, the new Batis API is more primitive.
There are no more automatic re-renderings as a result of asynchronous state changes.
This eliminates the need for the event system including the `Subject` class.
The listener function which is initially passed to the `Host` class is called instead.